refactor(billing): Extract MeteredCreditService and BillingSubscriptionUpdateService
- Create MeteredCreditService (~120 lines) to consolidate:
  - Metered pricing info lookups
  - Billing alert creation/recreation
  - Credit balance queries

- Create BillingSubscriptionUpdateService (~450 lines) to handle:
  - All subscription modification operations (plan, interval, seats, metered price)
  - Subscription scheduling logic

- Slim down BillingSubscriptionService from ~1205 to ~375 lines:
  - Keep CRUD operations
  - Keep trial management
  - Keep database sync
  - Delegate to new services for updates and alerts

- Update webhook services to use MeteredCreditService
- Rename workflow terminology to metered where appropriate
- Update and create tests for new services
