+++ packages/twenty-server/src/engine/core-modules/billing/billing.module.ts
@@ -17,16 +17,18 @@ import { BillingSubscriptionEntity } from 'src/engine/core-modules/billing/entit
 import { BillingRestApiExceptionFilter } from 'src/engine/core-modules/billing/filters/billing-api-exception.filter';
 import { BillingFeatureUsedListener } from 'src/engine/core-modules/billing/listeners/billing-feature-used.listener';
 import { BillingWorkspaceMemberListener } from 'src/engine/core-modules/billing/listeners/billing-workspace-member.listener';
+import { BillingCreditRolloverService } from 'src/engine/core-modules/billing/services/billing-credit-rollover.service';
 import { BillingPlanService } from 'src/engine/core-modules/billing/services/billing-plan.service';
 import { BillingPortalWorkspaceService } from 'src/engine/core-modules/billing/services/billing-portal.workspace-service';
 import { BillingPriceService } from 'src/engine/core-modules/billing/services/billing-price.service';
 import { BillingProductService } from 'src/engine/core-modules/billing/services/billing-product.service';
 import { BillingSubscriptionItemService } from 'src/engine/core-modules/billing/services/billing-subscription-item.service';
 import { BillingSubscriptionPhaseService } from 'src/engine/core-modules/billing/services/billing-subscription-phase.service';
+import { BillingSubscriptionUpdateService } from 'src/engine/core-modules/billing/services/billing-subscription-update.service';
 import { BillingSubscriptionService } from 'src/engine/core-modules/billing/services/billing-subscription.service';
 import { BillingUsageService } from 'src/engine/core-modules/billing/services/billing-usage.service';
-import { BillingCreditRolloverService } from 'src/engine/core-modules/billing/services/billing-credit-rollover.service';
 import { BillingService } from 'src/engine/core-modules/billing/services/billing.service';
+import { MeteredCreditService } from 'src/engine/core-modules/billing/services/metered-credit.service';
 import { StripeModule } from 'src/engine/core-modules/billing/stripe/stripe.module';
 import { WorkspaceDomainsModule } from 'src/engine/core-modules/domain/workspace-domains/workspace-domains.module';
 import { FeatureFlagEntity } from 'src/engine/core-modules/feature-flag/feature-flag.entity';
@@ -64,6 +66,7 @@ import { PermissionsModule } from 'src/engine/metadata-modules/permissions/permi
   ],
   providers: [
     BillingSubscriptionService,
+    BillingSubscriptionUpdateService,
     BillingSubscriptionItemService,
     BillingPortalWorkspaceService,
     BillingProductService,
@@ -80,14 +83,17 @@ import { PermissionsModule } from 'src/engine/metadata-modules/permissions/permi
     BillingUsageService,
     BillingPriceService,
     BillingCreditRolloverService,
+    MeteredCreditService,
   ],
   exports: [
     BillingSubscriptionService,
+    BillingSubscriptionUpdateService,
     BillingSubscriptionItemService,
     BillingPortalWorkspaceService,
     BillingService,
     BillingUsageService,
     BillingCreditRolloverService,
+    MeteredCreditService,
   ],
 })
 export class BillingModule {}
