+++ packages/twenty-server/src/engine/core-modules/billing/services/billing-subscription.service.ts
@@ -6,7 +6,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { differenceInDays } from 'date-fns';
 import {
   assertIsDefinedOrThrow,
-  assertUnreachable,
   findOrThrow,
   isDefined,
 } from 'twenty-shared/utils';
@@ -24,67 +23,43 @@ import {
   BillingException,
   BillingExceptionCode,
 } from 'src/engine/core-modules/billing/billing.exception';
-import { billingValidator } from 'src/engine/core-modules/billing/billing.validate';
 import { BillingCustomerEntity } from 'src/engine/core-modules/billing/entities/billing-customer.entity';
 import { BillingEntitlementEntity } from 'src/engine/core-modules/billing/entities/billing-entitlement.entity';
-import { BillingPriceEntity } from 'src/engine/core-modules/billing/entities/billing-price.entity';
 import { BillingSubscriptionItemEntity } from 'src/engine/core-modules/billing/entities/billing-subscription-item.entity';
 import { BillingSubscriptionEntity } from 'src/engine/core-modules/billing/entities/billing-subscription.entity';
 import { type BillingEntitlementKey } from 'src/engine/core-modules/billing/enums/billing-entitlement-key.enum';
-import { BillingPlanKey } from 'src/engine/core-modules/billing/enums/billing-plan-key.enum';
 import { BillingProductKey } from 'src/engine/core-modules/billing/enums/billing-product-key.enum';
-import { SubscriptionInterval } from 'src/engine/core-modules/billing/enums/billing-subscription-interval.enum';
 import { SubscriptionStatus } from 'src/engine/core-modules/billing/enums/billing-subscription-status.enum';
 import { BillingPlanService } from 'src/engine/core-modules/billing/services/billing-plan.service';
 import { BillingPriceService } from 'src/engine/core-modules/billing/services/billing-price.service';
-import { BillingProductService } from 'src/engine/core-modules/billing/services/billing-product.service';
-import { BillingSubscriptionPhaseService } from 'src/engine/core-modules/billing/services/billing-subscription-phase.service';
-import { StripeBillingAlertService } from 'src/engine/core-modules/billing/stripe/services/stripe-billing-alert.service';
-import { StripeCreditGrantService } from 'src/engine/core-modules/billing/stripe/services/stripe-credit-grant.service';
+import { MeteredCreditService } from 'src/engine/core-modules/billing/services/metered-credit.service';
 import { StripeCustomerService } from 'src/engine/core-modules/billing/stripe/services/stripe-customer.service';
 import { StripeSubscriptionScheduleService } from 'src/engine/core-modules/billing/stripe/services/stripe-subscription-schedule.service';
 import { StripeSubscriptionService } from 'src/engine/core-modules/billing/stripe/services/stripe-subscription.service';
-import {
-  type SubscriptionUpdate,
-  SubscriptionUpdateType,
-} from 'src/engine/core-modules/billing/types/billing-subscription-update.type';
-import { computeSubscriptionUpdateOptions } from 'src/engine/core-modules/billing/utils/compute-subscription-update-options.util';
-import { getCurrentLicensedBillingSubscriptionItemOrThrow } from 'src/engine/core-modules/billing/utils/get-licensed-billing-subscription-item-or-throw.util';
-import { getCurrentMeteredBillingSubscriptionItemOrThrow } from 'src/engine/core-modules/billing/utils/get-metered-billing-subscription-item-or-throw.util';
 import { getPlanKeyFromSubscription } from 'src/engine/core-modules/billing/utils/get-plan-key-from-subscription.util';
-import { getSubscriptionPricesFromSchedulePhase } from 'src/engine/core-modules/billing/utils/get-subscription-prices-from-schedule-phase.util';
 import { TwentyConfigService } from 'src/engine/core-modules/twenty-config/twenty-config.service';
 import { type WorkspaceEntity } from 'src/engine/core-modules/workspace/workspace.entity';
 
-type SubscriptionStripePrices = {
-  licensedPriceId: string;
-  seats: number;
-  meteredPriceId: string;
-};
 @Injectable()
 export class BillingSubscriptionService {
   protected readonly logger = new Logger(BillingSubscriptionService.name);
+
   constructor(
     private readonly stripeSubscriptionService: StripeSubscriptionService,
     private readonly billingPriceService: BillingPriceService,
     private readonly billingPlanService: BillingPlanService,
-    private readonly billingProductService: BillingProductService,
     @InjectRepository(BillingEntitlementEntity)
     private readonly billingEntitlementRepository: Repository<BillingEntitlementEntity>,
     @InjectRepository(BillingSubscriptionEntity)
     private readonly billingSubscriptionRepository: Repository<BillingSubscriptionEntity>,
     private readonly stripeCustomerService: StripeCustomerService,
     private readonly twentyConfigService: TwentyConfigService,
-    @InjectRepository(BillingPriceEntity)
-    private readonly billingPriceRepository: Repository<BillingPriceEntity>,
     @InjectRepository(BillingSubscriptionItemEntity)
     private readonly billingSubscriptionItemRepository: Repository<BillingSubscriptionItemEntity>,
     private readonly stripeSubscriptionScheduleService: StripeSubscriptionScheduleService,
-    private readonly billingSubscriptionPhaseService: BillingSubscriptionPhaseService,
     @InjectRepository(BillingCustomerEntity)
     private readonly billingCustomerRepository: Repository<BillingSubscriptionEntity>,
-    private readonly stripeBillingAlertService: StripeBillingAlertService,
-    private readonly stripeCreditGrantService: StripeCreditGrantService,
+    private readonly meteredCreditService: MeteredCreditService,
   ) {}
 
   async getBillingSubscriptions(workspaceId: string) {
@@ -217,68 +192,6 @@ export class BillingSubscriptionService {
     return entitlement.value;
   }
 
-  async changeMeteredPrice(
-    workspaceId: string,
-    meteredPriceId: string,
-  ): Promise<void> {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    const subscriptionUpdate = {
-      type: SubscriptionUpdateType.METERED_PRICE,
-      newMeteredPriceId: meteredPriceId,
-    } as const;
-
-    const isScheduledForPeriodEnd =
-      await this.shouldUpdateAtSubscriptionPeriodEnd(
-        billingSubscription,
-        subscriptionUpdate,
-      );
-
-    await this.updateSubscription(billingSubscription.id, subscriptionUpdate);
-
-    if (!isScheduledForPeriodEnd) {
-      await this.billingSubscriptionItemRepository.update(
-        { stripeSubscriptionId: billingSubscription.stripeSubscriptionId },
-        { hasReachedCurrentPeriodCap: false },
-      );
-
-      const newTierCap =
-        await this.getWorkflowTierCapFromPriceId(meteredPriceId);
-
-      const unitPriceCents =
-        await this.getWorkflowUnitPriceFromPriceId(meteredPriceId);
-
-      const creditBalance =
-        await this.stripeCreditGrantService.getCustomerCreditBalance(
-          billingSubscription.stripeCustomerId,
-          unitPriceCents,
-        );
-
-      await this.stripeBillingAlertService.createUsageThresholdAlertForCustomerMeter(
-        billingSubscription.stripeCustomerId,
-        newTierCap,
-        creditBalance,
-        billingSubscription.currentPeriodStart,
-      );
-    }
-  }
-
-  async cancelSwitchMeteredPrice(workspace: WorkspaceEntity): Promise<void> {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId: workspace.id },
-    );
-
-    const currentMeteredPrice =
-      getCurrentMeteredBillingSubscriptionItemOrThrow(billingSubscription);
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.METERED_PRICE,
-      newMeteredPriceId: currentMeteredPrice.stripePriceId,
-    });
-  }
-
   async endTrialPeriod(workspace: WorkspaceEntity) {
     const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
       { workspaceId: workspace.id },
@@ -312,7 +225,7 @@ export class BillingSubscriptionService {
       { hasReachedCurrentPeriodCap: false },
     );
 
-    await this.createBillingAlertForCustomer(
+    await this.meteredCreditService.recreateBillingAlertForCustomer(
       billingSubscription.stripeCustomerId,
     );
 
@@ -351,92 +264,19 @@ export class BillingSubscriptionService {
       },
     );
 
-    const workflowSubscriptionItem =
+    const meteredSubscriptionItem =
       billingSubscription.billingSubscriptionItems.find(
         (item) =>
           item.billingProduct.metadata.productKey ===
           BillingProductKey.WORKFLOW_NODE_EXECUTION,
       );
 
-    if (!workflowSubscriptionItem) {
+    if (!meteredSubscriptionItem) {
       throw new BillingException(
         'Workflow subscription item not found',
         BillingExceptionCode.BILLING_SUBSCRIPTION_ITEM_NOT_FOUND,
       );
     }
-  }
-
-  async cancelSwitchPlan(workspaceId: string) {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    const currentPlan =
-      getCurrentLicensedBillingSubscriptionItemOrThrow(billingSubscription)
-        .billingProduct?.metadata.planKey;
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.PLAN,
-      newPlan: currentPlan,
-    });
-  }
-
-  async cancelSwitchInterval(workspaceId: string) {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    const currentInterval = billingSubscription.interval;
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.INTERVAL,
-      newInterval: currentInterval,
-    });
-  }
-
-  async changeInterval(workspaceId: string) {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    const currentInterval = billingSubscription.interval;
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.INTERVAL,
-      newInterval:
-        currentInterval === SubscriptionInterval.Month
-          ? SubscriptionInterval.Year
-          : SubscriptionInterval.Month,
-    });
-  }
-
-  async changePlan(workspaceId: string) {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    const currentPlan =
-      getCurrentLicensedBillingSubscriptionItemOrThrow(billingSubscription)
-        .billingProduct?.metadata.planKey;
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.PLAN,
-      newPlan:
-        currentPlan === BillingPlanKey.ENTERPRISE
-          ? BillingPlanKey.PRO
-          : BillingPlanKey.ENTERPRISE,
-    });
-  }
-
-  async changeSeats(workspaceId: string, newSeats: number) {
-    const billingSubscription = await this.getCurrentBillingSubscriptionOrThrow(
-      { workspaceId },
-    );
-
-    await this.updateSubscription(billingSubscription.id, {
-      type: SubscriptionUpdateType.SEATS,
-      newSeats,
-    });
   }
 
   async syncSubscriptionToDatabase(
@@ -553,652 +393,5 @@ export class BillingSubscriptionService {
           : 'BILLING_FREE_WORKFLOW_CREDITS_FOR_TRIAL_PERIOD_WITHOUT_CREDIT_CARD',
       ),
     );
-  }
-
-  async getWorkflowTierCapForSubscription(
-    subscriptionId: string,
-  ): Promise<number> {
-    const subscription = await this.billingSubscriptionRepository.findOneOrFail(
-      {
-        where: { id: subscriptionId },
-        relations: [
-          'billingSubscriptionItems',
-          'billingSubscriptionItems.billingProduct',
-          'billingSubscriptionItems.billingProduct.billingPrices',
-        ],
-      },
-    );
-
-    const workflowItem = subscription.billingSubscriptionItems.find(
-      (item) =>
-        item.billingProduct.metadata.productKey ===
-        BillingProductKey.WORKFLOW_NODE_EXECUTION,
-    );
-
-    if (!isDefined(workflowItem)) {
-      throw new BillingException(
-        'Workflow subscription item not found',
-        BillingExceptionCode.BILLING_SUBSCRIPTION_ITEM_NOT_FOUND,
-      );
-    }
-
-    const matchingPrice = workflowItem.billingProduct.billingPrices.find(
-      (price) => price.stripePriceId === workflowItem.stripePriceId,
-    );
-
-    if (!isDefined(matchingPrice)) {
-      throw new BillingException(
-        `Cannot find price for product ${workflowItem.stripeProductId}`,
-        BillingExceptionCode.BILLING_PRICE_NOT_FOUND,
-      );
-    }
-
-    billingValidator.assertIsMeteredTiersSchemaOrThrow(matchingPrice.tiers);
-
-    return matchingPrice.tiers[0].up_to;
-  }
-
-  async getWorkflowTierCapFromPriceId(meteredPriceId: string): Promise<number> {
-    const price = await this.billingPriceRepository.findOneOrFail({
-      where: { stripePriceId: meteredPriceId },
-    });
-
-    billingValidator.assertIsMeteredTiersSchemaOrThrow(price.tiers);
-
-    return price.tiers[0].up_to;
-  }
-
-  async getWorkflowUnitPriceFromPriceId(
-    meteredPriceId: string,
-  ): Promise<number> {
-    const price = await this.billingPriceRepository.findOneOrFail({
-      where: { stripePriceId: meteredPriceId },
-    });
-
-    billingValidator.assertIsMeteredTiersSchemaOrThrow(price.tiers);
-
-    return Number(price.tiers[1].unit_amount_decimal);
-  }
-
-  async createBillingAlertForCustomer(
-    stripeCustomerId: string,
-    periodStart?: Date,
-  ): Promise<void> {
-    const subscription = await this.getCurrentBillingSubscription({
-      stripeCustomerId,
-    });
-
-    if (!isDefined(subscription)) {
-      this.logger.warn(
-        `Cannot create billing alert: subscription not found for stripeCustomerId ${stripeCustomerId}`,
-      );
-
-      return;
-    }
-
-    const workflowPricingInfo =
-      await this.getWorkflowPricingInfoForSubscription(subscription.id);
-
-    if (!isDefined(workflowPricingInfo)) {
-      this.logger.warn(
-        `Cannot create billing alert: workflow pricing info not found for subscription ${subscription.id}`,
-      );
-
-      return;
-    }
-
-    const creditBalance =
-      await this.stripeCreditGrantService.getCustomerCreditBalance(
-        stripeCustomerId,
-        workflowPricingInfo.unitPriceCents,
-      );
-
-    // Use the subscription's current period start if not provided
-    const effectivePeriodStart = periodStart ?? subscription.currentPeriodStart;
-
-    await this.stripeBillingAlertService.createUsageThresholdAlertForCustomerMeter(
-      stripeCustomerId,
-      workflowPricingInfo.tierCap,
-      creditBalance,
-      effectivePeriodStart,
-    );
-  }
-
-  async getWorkflowPricingInfoForSubscription(
-    subscriptionId: string,
-  ): Promise<{ tierCap: number; unitPriceCents: number } | null> {
-    const subscription = await this.billingSubscriptionRepository.findOne({
-      where: { id: subscriptionId },
-      relations: [
-        'billingSubscriptionItems',
-        'billingSubscriptionItems.billingProduct',
-        'billingSubscriptionItems.billingProduct.billingPrices',
-      ],
-    });
-
-    if (!isDefined(subscription)) {
-      return null;
-    }
-
-    const workflowItem = subscription.billingSubscriptionItems.find(
-      (item) =>
-        item.billingProduct.metadata.productKey ===
-        BillingProductKey.WORKFLOW_NODE_EXECUTION,
-    );
-
-    if (!isDefined(workflowItem)) {
-      return null;
-    }
-
-    const matchingPrice = workflowItem.billingProduct.billingPrices.find(
-      (price) => price.stripePriceId === workflowItem.stripePriceId,
-    );
-
-    if (!isDefined(matchingPrice)) {
-      return null;
-    }
-
-    if (!billingValidator.isMeteredTiersSchema(matchingPrice.tiers)) {
-      return null;
-    }
-
-    return {
-      tierCap: matchingPrice.tiers[0].up_to,
-      unitPriceCents: Number(matchingPrice.tiers[1].unit_amount_decimal),
-    };
-  }
-
-  private async runSubscriptionUpdate({
-    stripeSubscriptionId,
-    licensedStripeItemId,
-    meteredStripeItemId,
-    licensedStripePriceId,
-    meteredStripePriceId,
-    seats,
-    anchor,
-    proration,
-    metadata,
-  }: {
-    stripeSubscriptionId: string;
-    licensedStripeItemId: string;
-    meteredStripeItemId: string;
-    licensedStripePriceId: string;
-    meteredStripePriceId: string;
-    seats: number;
-    anchor?: Stripe.SubscriptionUpdateParams.BillingCycleAnchor;
-    proration?: Stripe.SubscriptionUpdateParams.ProrationBehavior;
-    metadata?: Record<string, string>;
-  }) {
-    return await this.stripeSubscriptionService.updateSubscription(
-      stripeSubscriptionId,
-      {
-        items: [
-          {
-            id: licensedStripeItemId,
-            price: licensedStripePriceId,
-            quantity: seats,
-          },
-          { id: meteredStripeItemId, price: meteredStripePriceId },
-        ],
-        ...(anchor ? { billing_cycle_anchor: anchor } : {}),
-        ...(proration ? { proration_behavior: proration } : {}),
-        ...(metadata ? { metadata } : {}),
-        billing_thresholds:
-          await this.billingPriceService.getBillingThresholdsByMeterPriceId(
-            meteredStripePriceId,
-          ),
-      },
-    );
-  }
-
-  async updateSubscription(
-    subscriptionId: string,
-    subscriptionUpdate: SubscriptionUpdate,
-  ): Promise<void> {
-    const subscription = await this.billingSubscriptionRepository.findOneOrFail(
-      {
-        where: { id: subscriptionId },
-        relations: [
-          'billingSubscriptionItems',
-          'billingSubscriptionItems.billingProduct',
-        ],
-      },
-    );
-
-    const licensedItem =
-      getCurrentLicensedBillingSubscriptionItemOrThrow(subscription);
-    const meteredItem =
-      getCurrentMeteredBillingSubscriptionItemOrThrow(subscription);
-
-    const toUpdateCurrentPrices = await this.computeSubscriptionPricesUpdate(
-      subscriptionUpdate,
-      {
-        licensedPriceId: licensedItem.stripePriceId,
-        meteredPriceId: meteredItem.stripePriceId,
-        seats: licensedItem.quantity,
-      },
-    );
-
-    const { schedule, currentPhase, nextPhase } =
-      await this.stripeSubscriptionScheduleService.loadSubscriptionSchedule(
-        subscription.stripeSubscriptionId,
-      );
-
-    const shouldUpdateAtSubscriptionPeriodEnd =
-      await this.shouldUpdateAtSubscriptionPeriodEnd(
-        subscription,
-        subscriptionUpdate,
-      );
-
-    if (shouldUpdateAtSubscriptionPeriodEnd) {
-      if (!isDefined(schedule)) {
-        const { schedule, currentPhase } =
-          await this.stripeSubscriptionScheduleService.createSubscriptionSchedule(
-            subscription.stripeSubscriptionId,
-          );
-
-        await this.runSubscriptionScheduleUpdate({
-          stripeScheduleId: schedule.id,
-          toUpdateCurrentPrices: undefined,
-          toUpdateNextPrices: toUpdateCurrentPrices,
-          currentPhase:
-            this.billingSubscriptionPhaseService.toPhaseUpdateParams(
-              currentPhase,
-            ),
-          subscriptionCurrentPeriodEnd: Math.floor(
-            subscription.currentPeriodEnd.getTime() / 1000,
-          ),
-        });
-      } else {
-        assertIsDefinedOrThrow(nextPhase);
-        assertIsDefinedOrThrow(currentPhase);
-
-        const toUpdateNextPrices = await this.computeSubscriptionPricesUpdate(
-          subscriptionUpdate,
-          getSubscriptionPricesFromSchedulePhase(nextPhase),
-        );
-
-        await this.runSubscriptionScheduleUpdate({
-          stripeScheduleId: schedule.id,
-          toUpdateNextPrices,
-          toUpdateCurrentPrices: undefined,
-          currentPhase:
-            this.billingSubscriptionPhaseService.toPhaseUpdateParams(
-              currentPhase,
-            ),
-          subscriptionCurrentPeriodEnd: Math.floor(
-            subscription.currentPeriodEnd.getTime() / 1000,
-          ),
-        });
-      }
-    } else {
-      const subscriptionOptions =
-        computeSubscriptionUpdateOptions(subscriptionUpdate);
-
-      await this.runSubscriptionUpdate({
-        stripeSubscriptionId: subscription.stripeSubscriptionId,
-        licensedStripeItemId: licensedItem.stripeSubscriptionItemId,
-        meteredStripeItemId: meteredItem.stripeSubscriptionItemId,
-        licensedStripePriceId: toUpdateCurrentPrices.licensedPriceId,
-        meteredStripePriceId: toUpdateCurrentPrices.meteredPriceId,
-        seats: toUpdateCurrentPrices.seats,
-        ...subscriptionOptions,
-      });
-
-      if (isDefined(nextPhase)) {
-        assertIsDefinedOrThrow(schedule);
-        const { currentPhase: refreshedCurrentPhase } =
-          await this.stripeSubscriptionScheduleService.loadSubscriptionSchedule(
-            subscription.stripeSubscriptionId,
-          );
-
-        assertIsDefinedOrThrow(refreshedCurrentPhase);
-
-        const nextPhasePrices =
-          getSubscriptionPricesFromSchedulePhase(nextPhase);
-        const toUpdateNextPrices = await this.computeSubscriptionPricesUpdate(
-          subscriptionUpdate,
-          nextPhasePrices,
-        );
-
-        await this.runSubscriptionScheduleUpdate({
-          stripeScheduleId: schedule.id,
-          toUpdateNextPrices,
-          toUpdateCurrentPrices: undefined, //subscription update causes schedule current phase update
-          currentPhase:
-            this.billingSubscriptionPhaseService.toPhaseUpdateParams(
-              refreshedCurrentPhase,
-            ),
-          subscriptionCurrentPeriodEnd: Math.floor(
-            subscription.currentPeriodEnd.getTime() / 1000,
-          ),
-        });
-      }
-    }
-
-    await this.syncSubscriptionToDatabase(
-      subscription.workspaceId,
-      subscription.stripeSubscriptionId,
-    );
-  }
-
-  private async runSubscriptionScheduleUpdate({
-    stripeScheduleId,
-    toUpdateNextPrices,
-    toUpdateCurrentPrices,
-    currentPhase,
-    subscriptionCurrentPeriodEnd,
-  }: {
-    stripeScheduleId: string;
-    toUpdateNextPrices: SubscriptionStripePrices;
-    toUpdateCurrentPrices: SubscriptionStripePrices | undefined;
-    currentPhase: Stripe.SubscriptionScheduleUpdateParams.Phase;
-    subscriptionCurrentPeriodEnd: number;
-  }) {
-    let toUpdateCurrentPhase: Stripe.SubscriptionScheduleUpdateParams.Phase = {
-      ...currentPhase,
-      end_date: subscriptionCurrentPeriodEnd,
-    };
-
-    if (isDefined(toUpdateCurrentPrices)) {
-      toUpdateCurrentPhase =
-        await this.billingSubscriptionPhaseService.buildPhaseUpdateParams({
-          licensedStripePriceId: toUpdateCurrentPrices.licensedPriceId,
-          seats: toUpdateCurrentPrices.seats,
-          meteredStripePriceId: toUpdateCurrentPrices.meteredPriceId,
-          endDate: subscriptionCurrentPeriodEnd,
-          startDate: currentPhase.start_date,
-        });
-    }
-
-    const toUpdateNextPhase =
-      await this.billingSubscriptionPhaseService.buildPhaseUpdateParams({
-        licensedStripePriceId: toUpdateNextPrices.licensedPriceId,
-        seats: toUpdateNextPrices.seats,
-        meteredStripePriceId: toUpdateNextPrices.meteredPriceId,
-        startDate: subscriptionCurrentPeriodEnd,
-        endDate: undefined,
-      });
-
-    if (
-      await this.billingSubscriptionPhaseService.isSamePhaseSignature(
-        toUpdateCurrentPhase,
-        toUpdateNextPhase,
-      )
-    ) {
-      return await this.stripeSubscriptionScheduleService.releaseSubscriptionSchedule(
-        stripeScheduleId,
-      );
-    }
-
-    return await this.stripeSubscriptionScheduleService.updateSchedule(
-      stripeScheduleId,
-      {
-        phases: [toUpdateCurrentPhase, toUpdateNextPhase],
-      },
-    );
-  }
-
-  private async shouldUpdateAtSubscriptionPeriodEnd(
-    subscription: BillingSubscriptionEntity,
-    update: SubscriptionUpdate,
-  ): Promise<boolean> {
-    switch (update.type) {
-      case SubscriptionUpdateType.PLAN: {
-        const currentPlan =
-          subscription.billingSubscriptionItems[0].billingProduct?.metadata
-            .planKey;
-
-        const isDowngrade =
-          currentPlan !== update.newPlan &&
-          update.newPlan === BillingPlanKey.PRO;
-
-        return isDowngrade;
-      }
-      case SubscriptionUpdateType.METERED_PRICE: {
-        const currentMeteredPriceId =
-          subscription.billingSubscriptionItems.find(
-            (item) => item.quantity == null,
-          )?.stripePriceId;
-
-        assertIsDefinedOrThrow(currentMeteredPriceId);
-        const currentMeteredPrice =
-          await this.billingPriceRepository.findOneOrFail({
-            where: { stripePriceId: currentMeteredPriceId },
-            relations: ['billingProduct'],
-          });
-        const newMeteredPrice = await this.billingPriceRepository.findOneOrFail(
-          {
-            where: { stripePriceId: update.newMeteredPriceId },
-            relations: ['billingProduct'],
-          },
-        );
-
-        billingValidator.assertIsMeteredPrice(currentMeteredPrice);
-        billingValidator.assertIsMeteredPrice(newMeteredPrice);
-
-        const currentMeteredCap = currentMeteredPrice.tiers[0].up_to;
-        const newMeteredCap = newMeteredPrice.tiers[0].up_to;
-
-        const isDowngrade = currentMeteredCap > newMeteredCap;
-
-        return isDowngrade;
-      }
-      case SubscriptionUpdateType.SEATS:
-        return false;
-      case SubscriptionUpdateType.INTERVAL: {
-        const currentInterval = subscription.interval;
-        const isDowngrade =
-          currentInterval !== update.newInterval &&
-          update.newInterval === SubscriptionInterval.Month;
-
-        return isDowngrade;
-      }
-      default: {
-        return assertUnreachable(
-          update,
-          'Should never occur, add validator for new subscription update type',
-        );
-      }
-    }
-  }
-
-  async computeSubscriptionPricesUpdate(
-    update: SubscriptionUpdate,
-    currentPrices: SubscriptionStripePrices,
-  ): Promise<SubscriptionStripePrices> {
-    switch (update.type) {
-      case SubscriptionUpdateType.PLAN:
-        return await this.computeSubscriptionPricesUpdateByPlan(
-          update.newPlan,
-          currentPrices,
-        );
-      case SubscriptionUpdateType.METERED_PRICE:
-        return await this.computeSubscriptionPricesUpdateByMeteredPrice(
-          update.newMeteredPriceId,
-          currentPrices,
-        );
-      case SubscriptionUpdateType.SEATS:
-        return this.computeSubscriptionPricesUpdateBySeats(
-          update.newSeats,
-          currentPrices,
-        );
-      case SubscriptionUpdateType.INTERVAL:
-        return await this.computeSubscriptionPricesUpdateByInterval(
-          update.newInterval,
-          currentPrices,
-        );
-    }
-  }
-
-  private computeSubscriptionPricesUpdateBySeats(
-    newSeats: number,
-    currentPrices: SubscriptionStripePrices,
-  ): SubscriptionStripePrices {
-    return {
-      ...currentPrices,
-      seats: newSeats,
-    };
-  }
-
-  private async computeSubscriptionPricesUpdateByMeteredPrice(
-    newMeteredPriceId: string,
-    currentPrices: SubscriptionStripePrices,
-  ): Promise<SubscriptionStripePrices> {
-    const currentLicensedPrice =
-      await this.billingPriceRepository.findOneOrFail({
-        where: { stripePriceId: currentPrices.licensedPriceId },
-        relations: ['billingProduct'],
-      });
-    const currentInterval = currentLicensedPrice.interval;
-    const currentPlanKey =
-      currentLicensedPrice.billingProduct?.metadata.planKey;
-
-    assertIsDefinedOrThrow(currentPlanKey);
-
-    const newMeteredPrice = await this.billingPriceRepository.findOneOrFail({
-      where: { stripePriceId: newMeteredPriceId },
-      relations: ['billingProduct'],
-    });
-
-    billingValidator.assertIsMeteredPrice(newMeteredPrice);
-
-    const newInterval = newMeteredPrice.interval;
-    const newPlanKey = newMeteredPrice.billingProduct?.metadata.planKey;
-
-    if (newInterval === currentInterval && currentPlanKey === newPlanKey) {
-      return {
-        ...currentPrices,
-        meteredPriceId: newMeteredPriceId,
-      };
-    }
-
-    const newEquivalentMeteredPrice =
-      await this.billingPriceService.findEquivalentMeteredPrice({
-        meteredPrice: newMeteredPrice,
-        targetInterval: currentInterval,
-        targetPlanKey: currentPlanKey,
-        hasSameInterval: newInterval === currentInterval,
-        hasSamePlanKey: currentPlanKey === newPlanKey,
-      });
-
-    return {
-      ...currentPrices,
-      meteredPriceId: newEquivalentMeteredPrice.stripePriceId,
-    };
-  }
-
-  private async computeSubscriptionPricesUpdateByPlan(
-    newPlan: BillingPlanKey,
-    currentPrices: SubscriptionStripePrices,
-  ): Promise<SubscriptionStripePrices> {
-    const currentLicensedPrice =
-      await this.billingPriceRepository.findOneOrFail({
-        where: { stripePriceId: currentPrices.licensedPriceId },
-        relations: ['billingProduct'],
-      });
-
-    const currentInterval = currentLicensedPrice.interval;
-    const currentPlanKey =
-      currentLicensedPrice.billingProduct?.metadata.planKey;
-
-    assertIsDefinedOrThrow(currentPlanKey);
-
-    if (currentPlanKey === newPlan) {
-      return currentPrices;
-    }
-
-    const billingPricesPerPlanAndIntervalArray =
-      await this.billingProductService.getProductPrices({
-        interval: currentInterval,
-        planKey: newPlan,
-      });
-
-    const targetLicensedPrice = findOrThrow(
-      billingPricesPerPlanAndIntervalArray,
-      ({ billingProduct }) =>
-        billingProduct?.metadata.productKey === BillingProductKey.BASE_PRODUCT,
-    );
-
-    const currentMeteredPrice = await this.billingPriceRepository.findOneOrFail(
-      {
-        where: { stripePriceId: currentPrices.meteredPriceId },
-        relations: ['billingProduct'],
-      },
-    );
-
-    billingValidator.assertIsMeteredPrice(currentMeteredPrice);
-
-    const targetMeteredPrice =
-      await this.billingPriceService.findEquivalentMeteredPrice({
-        meteredPrice: currentMeteredPrice,
-        targetInterval: currentInterval,
-        targetPlanKey: newPlan,
-        hasSameInterval: true,
-        hasSamePlanKey: false,
-      });
-
-    return {
-      ...currentPrices,
-      licensedPriceId: targetLicensedPrice.stripePriceId,
-      meteredPriceId: targetMeteredPrice.stripePriceId,
-    };
-  }
-
-  private async computeSubscriptionPricesUpdateByInterval(
-    newInterval: SubscriptionInterval,
-    currentPrices: SubscriptionStripePrices,
-  ): Promise<SubscriptionStripePrices> {
-    const currentLicensedPrice =
-      await this.billingPriceRepository.findOneOrFail({
-        where: { stripePriceId: currentPrices.licensedPriceId },
-        relations: ['billingProduct'],
-      });
-
-    const currentInterval = currentLicensedPrice.interval;
-    const currentPlanKey =
-      currentLicensedPrice.billingProduct?.metadata.planKey;
-
-    assertIsDefinedOrThrow(currentPlanKey);
-
-    if (currentInterval === newInterval) {
-      return currentPrices;
-    }
-
-    const billingPricesPerPlanAndIntervalArray =
-      await this.billingProductService.getProductPrices({
-        interval: newInterval,
-        planKey: currentPlanKey,
-      });
-
-    const targetLicensedPrice = findOrThrow(
-      billingPricesPerPlanAndIntervalArray,
-      ({ billingProduct }) =>
-        billingProduct?.metadata.productKey === BillingProductKey.BASE_PRODUCT,
-    );
-
-    const currentMeteredPrice = await this.billingPriceRepository.findOneOrFail(
-      {
-        where: { stripePriceId: currentPrices.meteredPriceId },
-        relations: ['billingProduct'],
-      },
-    );
-
-    billingValidator.assertIsMeteredPrice(currentMeteredPrice);
-
-    const targetMeteredPrice =
-      await this.billingPriceService.findEquivalentMeteredPrice({
-        meteredPrice: currentMeteredPrice,
-        targetInterval: newInterval,
-        targetPlanKey: currentPlanKey,
-        hasSameInterval: false,
-        hasSamePlanKey: true,
-      });
-
-    return {
-      ...currentPrices,
-      licensedPriceId: targetLicensedPrice.stripePriceId,
-      meteredPriceId: targetMeteredPrice.stripePriceId,
-    };
   }
 }
